requirements.txt
airiot_python_sdk==4.0.2
jsons==1.6.3
requests==2.31.0
urllib3==2.0.7
setuptools~=68.2.2
PyYAML==6.0.1
paho-mqtt==1.6.1
grpcio==1.59.0
grpcio-tools==1.59.0
confluent-kafka==2.3.0
cacheout==0.14.1